Modify galaxy role; add roles dir to repo
We are working on a previous version of the ansible-galaxy role due to
issues with the latest version (breaks the image build).
However, there is at least one modification in the latest version that
we need to imcorporate into our build process (see change to
tasks/download.yml). Until we have resolved the issues with the new
role, I'm making such updates manually, and I'm adding our current role files to
the repo.